dev/requirements.txt
flask
gunicorn
google-cloud-logging